Added documentation and updated README
Also moved URL resolver to it's
own module.
