+++ Cargo.toml
@@ -5,5 +5,4 @@ members = [
     "pretend-codegen",
     "pretend-isahc",
     "pretend-reqwest",
-    "pretend-testclient",
 ]
