--- pretend-testclient/Cargo.toml
@@ -1,8 +1,0 @@
-[package]
-name = "pretend-testclient"
-version = "0.1.0"
-authors = ["Lucien XU <sfietkonstantin@free.fr>"]
-edition = "2018"
-
-[dependencies]
-pretend = { path = "../pretend" }
